backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
colorama==0.4.6
Django==5.1.5
django-cors-headers==4.6.0
django-ninja==1.3.0
gunicorn==23.0.0
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pydantic==2.10.5
pydantic_core==2.27.2
pytest==8.3.4
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2